Update.
Fixed KeyDown handling. Now the GameMainControl itself handles the events.
Added paint handling for GameMainControl.
Fixed the bug that GameMainControl would crash if no GameMain is attached to it.
Added EndLife to GameMain. This was necessary as the life of it would continue after the form was closed.
Other minor changes.